fix: profiling图表展示未开启错误 --Bug=129537093 # Reviewed, transaction id: 16371
